requirements.txt
albumentations==1.1.0
numpy==1.21.5
opencv-python==4.5.4.60
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
wandb==0.12.9
# (optional)